requirements.txt
torch
torchvision
opencv-python
matplotlib
lime
shap
numpy
pillow
scikit-image
mlflow
